noir_ssa_executor
Module compiler

Functions

compile_from_artifacts

Compiles the given FunctionBuilder into a CompiledProgram its taken from noirc_driver::compile_no_check, but modified to accept ArtifactsAndWarnings

compile_from_ssa 🔒

convert_generated_acir_into_circuit_without_signature 🔒

Converts the generated ACIR into a circuit artifact its taken from noirc_evaluator::ssa::convert_generated_acir_into_circuit, but modified not to use signature in initial function signature used to split public, private and return witnesses but now we don’t need it, because we don’t have any public inputs in SSA context, so we just use all witnesses as private inputs and return witnesses

create_program 🔒

Creates a program artifact from the given FunctionBuilder its taken from noirc_evaluator::ssa::create_program, but modified to accept FunctionBuilder

evaluator_options

optimize_ssa_into_acir

Optimizes the given SSA into ACIR