noir_ssa_executor::runner
Function execute

fn execute<B: BlackBoxFunctionSolver<FieldElement> + Default>(
    program: &Program<FieldElement>,
    initial_witness: WitnessMap<FieldElement>,
) -> Result<WitnessStack<FieldElement>, NargoError<FieldElement>>

Low level function to execute the given program with the given initial witness It uses nargo execute_program to run the program